Config sanity check; Alignment mode; Updates for the display unit; Better default config;

+++ location.h
@@ -4,9 +4,9 @@
 
 // altitude, latitude, longitude
 struct ObserverPosition {
-	float altitude;
-	float latitude;
-	float longitude;
+	double altitude;
+	double latitude;
+	double longitude;
 };
 
 struct RaDecPosition {
